refactors modules; creates bin dir for start-up scripts
