--- apollo/controller/__init__.py
@@ -1,1 +1,0 @@
-from .controller import Controller
