--- apollo/driver/__init__.py
@@ -1,1 +1,0 @@
-from .motor_controller import MotorController
