--- apollo/session/__init__.py
@@ -1,1 +1,0 @@
-from .session import Session
